resizer Service working + added to cron
