+++ app.js
@@ -1,11 +1,9 @@
 const express = require('express')
 const cors = require('cors')
 const multer = require( 'multer')
-const sharp = require('sharp')
 const Datastore = require('nedb')
 const { v4: uuidv4 } = require('uuid');
 const fs = require('fs');
-const path = require('path');
 const sizeOf = require('image-size');
 const jobResizer= require('./crons/cronResizer');
 
@@ -15,61 +13,12 @@ app.use(express.json())
 const PORT = 3000
 const md5File = require('md5-file')
 
-jobResizer.arrancarJob()
-
+// Configuracion bases de datos en memoria:
 const dbTasks = new Datastore({filename: __dirname + '/db/historicoTareas.dat', autoload: true});
 const dbImages = new Datastore({filename: __dirname + '/db/historicoImagenes.dat', autoload: true});
 
-// funcion para redimensionar ficheros con la dependencia sharp
-const renderizador = async (uuid, filePath, fileName, size, nombreSinExtension) =>{
-    // redimensionado imagen
-    fileresult= await sharp (filePath)
-    .resize(size)
-    .toFile("output/"+nombreSinExtension+"/"+size+"/"+fileName )
-    
-    // calcular MD5 fichero resultante
-    const hash= await md5File("output/"+nombreSinExtension+"/"+size+"/"+fileName)
-
-    // renombrar fichero con el MD5
-    const oldPath = path.join(__dirname, "output/",nombreSinExtension+"/"+size+"/"+fileName)  
-    const newPath = path.join(__dirname, "output",nombreSinExtension,size.toString(),hash + ".jpg")
-    
-    fs.renameSync(oldPath, newPath) 
-    
-    var file= {path: newPath, name: hash + ".jpg", md5: hash, fileresult:fileresult};
-
-    // cambiar estado tarea
-    dbTasks.update({_id: uuid}, {$set: {
-                                        nombre: file.name, 
-                                        path: file.path,
-                                        status: "finished",
-                                        lastUpdate:new Date().toISOString(),
-                                        md5: file.md5}
-                                    }, 
-                                {}, 
-                                function(err, num) {
-                                    if (err) {
-                                        console.error(err);
-                                        return;
-                                    }
-                                });
-    
-    console.log("Task finished: file already resized to "+ size +" " +uuid)
-        // Guardar imagen final
-    dbImages.insert({   timestamp: new Date().toISOString(),
-        md5: file.md5,
-        path: file.path,
-        width: fileresult.width,
-        height: fileresult.height
-        }, function(err, record) {
-            if (err) {
-                console.error(err);
-                return;
-            }
-        return record
-    });
-
-}
+//Sentencia para arrancar el job de procesar las imagenes encoladas cada X tiempo
+jobResizer.arrancarJob(dbTasks, dbImages)
 
 // configuracion para guardar la imagen enviada a traves del formdata gracias a la libreria multer
 const storage = multer.diskStorage ({
@@ -81,6 +30,10 @@ const storage = multer.diskStorage ({
     }
 })
 
+// middleware para guardar imagenes enviadas a traves de un form-data
+const upload = multer({ storage})
+
+// si no existen... creacion de los directorios donde se guardan las imagenes originales y de salida
 if (!fs.existsSync("Images")){
     fs.mkdirSync("Images");
 }
@@ -88,14 +41,10 @@ if (!fs.existsSync("output")){
     fs.mkdirSync("output");
 }
 
-// middleware para guardar imagenes enviadas a traves de un form-data
-const upload = multer({ storage})
-
-
 //endpoint para solicitar el redimensionado de una imagen
 app.post('/task', upload.single('file'), async (req, res) => {
     
-    console.log('llamada a endpoint de renderizado ->')
+    console.log('llamada a endpoint de encolado de tarea ->')
     const uuid800=uuidv4()
     const uuid1024=uuidv4()
     
@@ -122,16 +71,19 @@ app.post('/task', upload.single('file'), async (req, res) => {
     });
 
     var datenow=new Date().toISOString()
-    var docs = [
-        {_id: uuid800, status: "pending", imagenOriginal: req.file.filename, resolucion: "800", timestamp: datenow, lastUpdate: datenow},
-        {_id: uuid1024, status: "pending", imagenOriginal: req.file.filename, resolucion: "1024", timestamp: datenow, lastUpdate: datenow}
-    ];
-    
+
     nombreSinExtension= req.file.filename.slice(0, -4)
+
+    // creacion de directorios Output de la imagen que se va a generar
     fs.mkdirSync("output/"+nombreSinExtension);
     fs.mkdirSync("output/"+nombreSinExtension+"/800");
     fs.mkdirSync("output/"+nombreSinExtension+"/1024");
 
+
+    var docs = [
+        {_id: uuid800, status: "pending", path: "output/"+nombreSinExtension+"/800", imagenOriginal: req.file.path, resolucion: "800", timestamp: datenow, lastUpdate: datenow},
+        {_id: uuid1024, status: "pending", path: "output/"+nombreSinExtension+"/1024", imagenOriginal: req.file.path, resolucion: "1024", timestamp: datenow, lastUpdate: datenow}
+    ];
 
     // creacion registros de tareas encoladas en base de datos
     dbTasks.insert(docs, function(err, record) {
@@ -142,8 +94,8 @@ app.post('/task', upload.single('file'), async (req, res) => {
         return record
     });
 
-    renderizador(uuid1024, req.file.path, req.file.filename, 1024, nombreSinExtension)
-    renderizador(uuid800, req.file.path, req.file.filename, 800, nombreSinExtension)
+    // renderizador(uuid1024, req.file.path, req.file.filename, 1024, nombreSinExtension)
+    // renderizador(uuid800, req.file.path, req.file.filename, 800, nombreSinExtension)
     
 })
